feat: add generation reconciliation step and sync ready condition logic
